requirements.txt
# Core dependencies for Context Tools MCP Server
mcp>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
google-genai>=1.0.0
rapidfuzz>=3.0.0

# Development dependencies (install with: pip install -r requirements-dev.txt)
